refactor: 회원 API 개선 및 패키지 구조 정리
  - 프로필 이미지 업데이트 로직 간소화
  - getMyGroups 빈 리스트 가드 추가 (SQL IN () 에러 방지)
  - DTO 패키지 이동: controller/dto → dto/request, dto/respon
